requirements.txt
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
torch>=1.7.0
torchvision>=0.8.1
pandas>=1.1.4
seaborn>=0.11.0
pytz~=2022.1
gdown>=4.6.4
streamlit==1.26.0
